+++ ouzel/graphics/metal/tvos/MetalRenderDeviceTVOS.mm
@@ -43,7 +43,7 @@ namespace ouzel
         }
 
         void MetalRenderDeviceTVOS::init(Window* newWindow,
-                                         const Size2<float>& newSize,
+                                         const Size2<uint32_t>& newSize,
                                          uint32_t newSampleCount,
                                          Texture::Filter newTextureFilter,
                                          uint32_t newMaxAnisotropy,
