+++ ouzel/graphics/opengl/android/OGLRenderDeviceAndroid.cpp
@@ -76,7 +76,7 @@ namespace ouzel
         }
 
         void OGLRenderDeviceAndroid::init(Window* newWindow,
-                                          const Size2<float>&,
+                                          const Size2<uint32_t>&,
                                           uint32_t newSampleCount,
                                           Texture::Filter newTextureFilter,
                                           uint32_t newMaxAnisotropy,
@@ -170,8 +170,8 @@ namespace ouzel
             frameBufferWidth = surfaceWidth;
             frameBufferHeight = surfaceHeight;
 
-            Size2<float> backBufferSize = Size2<float>(static_cast<float>(frameBufferWidth),
-                                         static_cast<float>(frameBufferHeight));
+            Size2<uint32_t> backBufferSize = Size2<uint32_t>(static_cast<uint32_t>(frameBufferWidth),
+                                                             static_cast<uint32_t>(frameBufferHeight));
 
             OGLRenderDevice::init(newWindow,
                                   backBufferSize,
