+++ ouzel/graphics/opengl/windows/OGLRenderDeviceWin.cpp
@@ -150,7 +150,7 @@ namespace ouzel
         }
 
         void OGLRenderDeviceWin::init(Window* newWindow,
-                                      const Size2<float>& newSize,
+                                      const Size2<uint32_t>& newSize,
                                       uint32_t newSampleCount,
                                       Texture::Filter newTextureFilter,
                                       uint32_t newMaxAnisotropy,
